requirements.txt
asgiref==3.5.0
attrs==21.4.0
click==8.1.2
Django==3.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-extensions==3.1.5
djangorestframework==3.13.1
flake8==4.0.1
h11==0.13.0
iniconfig==1.1.1
mccabe==0.6.1
model-bakery==1.4.0
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
pytest==7.1.1
pytest-django==4.5.2
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
uvicorn==0.17.6
